HRM now hot swapable
git-svn-id: https://oceanai.mit.edu/svn/moos-ivp-aquaticus-oai/trunk@1552 34ae1d1b-fdb2-4983-80a9-df12a7f50355

+++ src/iZephyrHRM/ZephyrHRM.h
@@ -67,7 +67,7 @@ class ZephyrHRM : public AppCastingMOOSApp
    int m_bt_channel;
    int m_packet_num;
    int m_life_sign_c;
-   float m_last_life_sign_time;
+   double m_last_life_sign_time;
    static const int m_bt_data_buf_size = 1024;
    unsigned char m_bt_data_buf[m_bt_data_buf_size];
